The folder containing a precise multitude of the considerable amount of app icons we are currently in possession of. Our design team has spent many hours curating these icons from a singular source. Themselves. Having many custom app icons introduces an element of customizability. This causes the user to enjoy using the app more, because they feel like it's more personal. This, along with our other options, creates a positively exquisite user experience, leading to high reviews in the App Store (something we do not currently employ), with positive side effects including but not limited to happiness.
The folder containing a precise multitude of the considerable amount of app icons we are currently in possession of. Our design team has spent many hours curating these icons from a singular source. Themselves. Having many custom app icons introduces an element of customizability. This causes the user to enjoy using the app more, because they feel like it's more personal. This, along with our other options, creates a positively exquisite user experience, leading to high reviews in the App Store (something we do not currently employ), with positive side effects including but not limited to happiness.